⚡ :: UPDATE

+++ src/main/java/com/example/loginpractice/entity/diary/DiaryEntity.java
@@ -1,11 +1,8 @@
 package com.example.loginpractice.entity.diary;
 
 import lombok.*;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
-import java.util.Collection;
 
 @Entity
 @NoArgsConstructor
@@ -18,7 +15,7 @@ public class DiaryEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, length = 20)
     private String title;
